daemon/requirements.txt
beautifulsoup4
svglib
reportlab
Pillow
python-barcode
fpdf
svglib
